Maze-navigating robotic car
Centered around a raspberry pi, this project uses python programs to get data from different sensors, measuring distances in front of and beside the car, the light hitting the car, and the temperature around the car. Using this data, the pi drives four motors to navigate a maze. The maze designed for the class featured a dark hallway and a hot hallway, both of which were to be avoided, hence the light and temperature sensors. The frontmost ultrasonic distance sensor is mounted on a servo drivenby th pi, in order to measure the distance between the car and a slightly larger area in front of it than a fixed sensor would.